config: Fix initialization of 'unsigned int' from 'void *' makes integer from pointer without a cast

+++ config.def.h
@@ -41,7 +41,7 @@ static const char *const autostart[] = {
 /* NOTE: ALWAYS keep a rule declared even if you don't use rules (e.g leave at least one example) */
 static const Rule rules[] = {
 	/* app_id             title       tags mask     isfloating   monitor */
-  { NULL,     "float",    NULL,            1,           -1 }
+  { NULL,     "float",    (unsigned int) NULL,            1,           -1 }
 };
 
 /* layout(s) */
